changing mathproblems.java
making it shorter and simpler with functions

+++ app/src/main/java/_training/Return.java
@@ -1,16 +1,16 @@
 package _training;
 
 public class Return {
-    static float square(float num){
+    static int square(int num){
         return num*num;
     }
     public static void main(String[] args){
-        float square5 = square(5);
-        float square2 = square(2);
+        int square5 = square(5);
+        int square2 = square(2);
         System.out.println(square5 + square2);
 
-        float square10 = square(10);
-        float square6 = square(6);
+        int square10 = square(10);
+        int square6 = square(6);
         System.out.println(square10 * square6);
